Fix error in certificate handling (#27)
Fixed code that fails with the AWS v3 provider for infrastructure definitions
meeting the following conditions:

```
inputs = {
  ...
  create_route53_record = false
  create_acm_cert       = false
}
```

The problem was a `for_each` expression that didn't allow for this particular
combination of inputs. This use case didn't exist in the test account, and
hence wasn't detected during provider upgrade there.

Change module reference to canonical GitHub URL so that GitHub Actions
can import the `terraform-aws-util` repo, which is now public.